third_party/blink/web_tests/virtual/isolate-sandboxed-iframes/external/wpt/html/infrastructure/urls/terminology-0/document-base-url-about-srcdoc.https.window-expected.txt

This is a testharness.js-based test.
PASS allow-same-origin
FAIL disallow-same-origin assert_equals: expected "https://web-platform.test:8444/html/infrastructure/urls/terminology-0/document-base-url-about-srcdoc.https.window.html" but got "about:srcdoc"
Harness: the test ran to completion.
